+++ crawler/src/instances/external.rs
@@ -1,2 +1,0 @@
-#[path = "../../../../apps/garbage/crawler/src/instances/garbage.rs"]
-pub mod garbage;